+++ src/stopwatch.rs
@@ -144,10 +144,10 @@ impl Stopwatch {
     /// # Examples
     ///
     /// ```
-    /// # use libsw::{Error, Stopwatch};
+    /// # use libsw::{Result, Stopwatch};
     /// # use core::time::Duration;
     /// # use std::thread;
-    /// # fn main() -> Result<(), Error> {
+    /// # fn main() -> Result<()> {
     /// let sw = Stopwatch::new_started();
     /// thread::sleep(Duration::from_millis(100));
     /// assert!(sw.elapsed() >= Duration::from_millis(100));
@@ -205,7 +205,7 @@ impl Stopwatch {
     /// assert!(then != now);
     /// ```
     #[inline]
-    pub fn start(&mut self) -> Result<(), Error> {
+    pub fn start(&mut self) -> Result<()> {
         self.start_at(Instant::now())
     }
 
@@ -232,7 +232,7 @@ impl Stopwatch {
     /// assert!(then == now);
     /// ```
     #[inline]
-    pub fn stop(&mut self) -> Result<(), Error> {
+    pub fn stop(&mut self) -> Result<()> {
         self.stop_at(Instant::now())
     }
 
@@ -246,11 +246,11 @@ impl Stopwatch {
     /// # Examples
     ///
     /// ```
-    /// # use libsw::{Error, Stopwatch};
+    /// # use libsw::{Result, Stopwatch};
     /// # use core::time::Duration;
     /// # use std::thread;
     /// # use std::time::Instant;
-    /// # fn main() -> Result<(), Error> {
+    /// # fn main() -> Result<()> {
     /// let mut sw_1 = Stopwatch::new();
     /// let mut sw_2 = Stopwatch::new();
     ///
@@ -267,7 +267,7 @@ impl Stopwatch {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn start_at(&mut self, anchor: Instant) -> Result<(), Error> {
+    pub fn start_at(&mut self, anchor: Instant) -> Result<()> {
         if self.is_running() {
             Err(Error::AlreadyStarted)
         } else {
@@ -288,11 +288,11 @@ impl Stopwatch {
     /// # Examples
     ///
     /// ```
-    /// # use libsw::{Error, Stopwatch};
+    /// # use libsw::{Result, Stopwatch};
     /// # use core::time::Duration;
     /// # use std::thread;
     /// # use std::time::Instant;
-    /// # fn main() -> Result<(), Error> {
+    /// # fn main() -> Result<()> {
     /// let mut sw_1 = Stopwatch::new_started();
     /// let mut sw_2 = sw_1.clone();
     /// let stop = Instant::now();
@@ -302,7 +302,7 @@ impl Stopwatch {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn stop_at(&mut self, anchor: Instant) -> Result<(), Error> {
+    pub fn stop_at(&mut self, anchor: Instant) -> Result<()> {
         if let Some(start) = self.start {
             *self += anchor.saturating_duration_since(start);
             self.start = None;
@@ -341,7 +341,7 @@ impl Stopwatch {
     ///
     /// Returns [`AlreadyStarted`](Error::AlreadyStarted) if the stopwatch is
     /// running.
-    pub fn guard(&mut self) -> Result<Guard<'_>, Error> {
+    pub fn guard(&mut self) -> Result<Guard<'_>> {
         self.start()?;
         Ok(Guard { inner: self })
     }
@@ -357,7 +357,7 @@ impl Stopwatch {
     ///
     /// Returns [`AlreadyStarted`](Error::AlreadyStarted) if the stopwatch is
     /// running.
-    pub fn guard_at(&mut self, anchor: Instant) -> Result<Guard<'_>, Error> {
+    pub fn guard_at(&mut self, anchor: Instant) -> Result<Guard<'_>> {
         self.start_at(anchor)?;
         Ok(Guard { inner: self })
     }
@@ -582,13 +582,16 @@ impl Hash for Stopwatch {
     }
 }
 
+/// Alias for `Result<T, Error>`.
+pub type Result<T> = std::result::Result<T, Error>;
+
 /// Errors associated with a [`Stopwatch`].
 ///
 /// # Examples
 ///
 /// ```
-/// # use libsw::{Error, Stopwatch};
-/// # fn main() -> Result<(), Error> {
+/// # use libsw::{Error, Result, Stopwatch};
+/// # fn main() -> Result<()> {
 /// let mut sw = Stopwatch::new_started();
 /// assert_eq!(sw.start(), Err(Error::AlreadyStarted));
 /// sw.stop()?;
@@ -605,7 +608,7 @@ pub enum Error {
 }
 
 impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> std::result::Result<(), fmt::Error> {
         f.write_str(match self {
             Self::AlreadyStarted => "stopwatch already started",
             Self::AlreadyStopped => "stopwatch already stopped",