+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [unreleased]
 * added `PartialEq` and `Eq` impl for `Stopwatch`
+* added `Result<T>` type, an alias for `Result<T, Error>`
 * updated documentation for `Stopwatch::from_raw`
 * minor docs improvements